Refactor of EvalExprWrapper
