+++ partiql-eval/src/eval/expr/mod.rs
@@ -1,4 +1,4 @@
-pub(crate) mod arg_check;
+pub(crate) mod eval_wrapper;
 pub(crate) mod pattern_match;
 
 mod coll;
